refactor: move logger to utils folder
Organize logger into src/utils/ for better code structure.
